Separate Repository module from Core module
