+++ DBMonitoringWindowApp/build.gradle
@@ -14,6 +14,7 @@ dependencies {
     implementation project(':Utils')
     implementation project(':Common')
     implementation project(':Core')
+    implementation project(':Repository')
     
 	// Quartz Scheduler
     implementation 'org.quartz-scheduler:quartz:2.3.2'
