+++ DBMonitoringWindowApp/src/main/java/root/javafx/Controller/HistoryMenuController.java
@@ -22,17 +22,17 @@ import root.core.domain.OSDiskUsage;
 import root.core.domain.TableSpaceUsage;
 import root.core.repository.constracts.DBCheckRepository;
 import root.core.repository.constracts.ServerMonitoringRepository;
-import root.core.repository.implement.DBCheckRepositoryImpl;
-import root.core.repository.implement.LinuxServerMonitoringRepository;
-import root.core.repository.implement.PropertyRepositoryImpl;
-import root.core.repository.implement.ReportFileRepo;
 import root.core.service.contracts.PropertyService;
-import root.core.service.implement.FilePropertyService;
 import root.core.usecase.constracts.DBCheckUsecase;
 import root.core.usecase.constracts.ServerMonitoringUsecase;
 import root.core.usecase.implement.DBCheckUsecaseImpl;
 import root.core.usecase.implement.ServerMonitoringUsecaseImpl;
 import root.javafx.utils.AlertUtils;
+import root.repository.implement.DBCheckRepositoryImpl;
+import root.repository.implement.LinuxServerMonitoringRepository;
+import root.repository.implement.PropertyRepositoryImpl;
+import root.repository.implement.ReportFileRepo;
+import root.service.implement.FilePropertyService;
 
 public class HistoryMenuController implements Initializable {
