+++ DBMonitoringWindowApp/src/main/java/root/javafx/Controller/MonitoringAPController.java
@@ -35,10 +35,7 @@ import javafx.scene.layout.HBox;
 import javafx.util.StringConverter;
 import root.core.domain.MonitoringResult;
 import root.core.domain.enums.RoundingDigits;
-import root.core.repository.implement.PropertyRepositoryImpl;
-import root.core.repository.implement.ReportFileRepo;
 import root.core.service.contracts.PropertyService;
-import root.core.service.implement.FilePropertyService;
 import root.core.usecase.constracts.ReportUsecase;
 import root.core.usecase.implement.ReportUsecaseImpl;
 import root.javafx.CustomView.MonitoringTableViewContainer;
@@ -46,6 +43,9 @@ import root.javafx.CustomView.dateCell.MonitoringHistoryDateCell;
 import root.javafx.CustomView.prequencyUI.PrequencyButton;
 import root.javafx.DI.DependencyInjection;
 import root.javafx.utils.AlertUtils;
+import root.repository.implement.PropertyRepositoryImpl;
+import root.repository.implement.ReportFileRepo;
+import root.service.implement.FilePropertyService;
 import root.utils.DateUtils;
 import root.utils.UnitUtils.FileSize;
