+++ DBMonitoringWindowApp/src/main/java/root/javafx/Controller/SettingMenuController.java
@@ -37,15 +37,15 @@ import root.core.domain.MonitoringYN.MonitoringTypeAndYN;
 import root.core.domain.enums.MonitoringType;
 import root.core.domain.enums.RoundingDigits;
 import root.core.domain.enums.UsageUIType;
-import root.core.repository.implement.PropertyRepositoryImpl;
 import root.core.service.contracts.PropertyService;
-import root.core.service.implement.FilePropertyService;
 import root.javafx.CustomView.ConnectionInfoVBox;
 import root.javafx.CustomView.DBConnInfoControl;
 import root.javafx.CustomView.MonitoringYNVBox;
 import root.javafx.CustomView.ServerConnInfoControl;
 import root.javafx.CustomView.dialogUI.CustomTextInputDialog;
 import root.javafx.utils.AlertUtils;
+import root.repository.implement.PropertyRepositoryImpl;
+import root.service.implement.FilePropertyService;
 import root.utils.UnitUtils.FileSize;
 
 @Slf4j
