+++ Repository/.gitignore
@@ -1,0 +1,2 @@
+/build/
+/bin/
