+++ Repository/src/main/java/root/repository/implement/PropertyRepositoryImpl.java
@@ -1,4 +1,4 @@
-package root.core.repository.implement;
+package root.repository.implement;
 
 import java.io.BufferedWriter;
 import java.io.File;
