+++ Repository/src/main/java/root/service/implement/FilePropertyService.java
@@ -1,4 +1,4 @@
-package root.core.service.implement;
+package root.service.implement;
 
 import java.util.ArrayList;
 import java.util.Arrays;
