+++ Repository/src/test/java/root/repository/implement/LinuxServerMonitoringRepositoryTest.java
@@ -1,4 +1,4 @@
-package root.core.repository.implement;
+package root.repository.implement;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
