+++ settings.gradle
@@ -3,6 +3,7 @@ rootProject.name = 'DBMonitoring'
 include ":Utils"
 include ":Common"
 include ":Core"
+include ":Repository"
 include ":DBMonitoringWindowApp"
 include ":DBMonitoringDownloaderApp"
 
@@ -10,5 +11,6 @@ include ":DBMonitoringDownloaderApp"
 project(":Utils").projectDir = file("Utils")
 project(":Common").projectDir = file("Common")
 project(":Core").projectDir = file("Core")
+project(":Repository").projectDir = file("Repository")
 project(":DBMonitoringWindowApp").projectDir = file("DBMonitoringWindowApp")
 project(":DBMonitoringDownloaderApp").projectDir = file("DBMonitoringDownloaderApp")
